* ci: add local verify script and Windows directml clippy gate

release.yml ships a --features directml Windows binary, but no PR check
compiled that feature, so an ort API rename only failed the release build
on main. verify.sh runs every CI gate locally, in CI order, under a
throwaway TOKENIX_HOME so it never touches the user's real state.


* fix(security): gate repo rule files, centralize state dir, harden persistence

A cloned repo could still steer tokenix without `tokenix trust`: repo
secret-rules and egress-rules loaded unconditionally, so an added `.+`
rule would blank every PostToolUse result. They now load only once the
repo is trusted and stay add-only. Memory notes are also checked against
the bundled secret rules, so bare AKIA/ghp_ tokens are refused.

Persisted views (recordings, unfiltered log, snapshots) are redacted, the
daemon status endpoint needs the capability token, re-read suppression is
scoped to the agent session, and filter generate asks before re-running.

store::global_dir() is now the single source of ~/.tokenix and honours an
absolute TOKENIX_HOME, so tests and verify.sh never write the real home.
find_project_root prefers an indexed directory over marker files: a
package.json in a home directory captured every marker-less repo below it.

search_similar scans vectors only and loads text for the top k
(122-130 ms -> 47-48 ms/query on a 30k x 768d bench). fastembed 4.9 -> 7
and ort rc.13 (models behind a Mutex, renamed EPs); nomic vectors are
bit-identical, so existing indexes stay valid. rustls bumped for
RUSTSEC-2026-0285. Dependency and code changes share embed.rs, so they
ship together.


* chore(deps): align fuzz crate on tree-sitter 0.27

The fuzz target compiles src/chunker.rs against its own lockfile, which
dependabot left on tree-sitter 0.26 when it bumped the main crate. In
0.27 child_count() returns u32, so the now-redundant `as u32` cast was
dropped from chunker.rs, and the 0.26 fuzz build broke (E0308 in
cflite_pr.yml). The fuzzer should exercise the parser version that ships.


* ci: type-check fuzz targets in verify.sh

The fuzz crate has its own lockfile and is only built by the
ClusterFuzzLite PR job, so a signature drift in chunker.rs passed every
local gate and failed only after push.


* ci: keep ort's download cache inside the cached target dir

ort-sys rc.13 extracts ONNX Runtime into a per-user cache and symlinks
DirectML.dll from there into target/. The Windows test job restores
target/ from actions/cache but not that directory, and the build script
does not rerun, so the test binary died with STATUS_DLL_NOT_FOUND on
every warm-cache run. The cache key changes so no entry holding dangling
links is restored.


* fix(index): stop perpetual inline refresh and cap git-planned file size

The inline refresh counted every dirty file the index did not know as new.
Empty, binary and sub-MIN_CHUNK_TOKENS files never get a row, so a single
untracked empty file made every retrieval command pay a ~80 ms refresh,
forever. It now asks indexer::stores_content, which runs the same
decode + chunk path the indexer uses to decide whether a row is written.

The git-incremental plan also skipped the max_file_bytes cap that the full
walk and restored_files apply, so a dirty file above the cap was indexed on
the incremental path only.


* test: end-to-end suites proving the core workflow and safety guarantees

homologation.sh mostly asserts "did not panic". These suites spawn the real
binary against a fixture repo and assert what a user relies on: exact
symbol locations, callers/callees, outline vs --symbol body, Read/Grep hook
decisions with a real index, fail-open without one, pack budget, exit code
kept by tokenix run, trust gate incl. revocation on edit, memory refusing
credentials, and edits searchable without re-indexing.

Every test gets its own git repo and TOKENIX_HOME, and hook_e2e now sets
TOKENIX_HOME too; it used to write into the developer's ~/.tokenix.


* docs: turn the README into a guide and add one guide per agent

The README read as a reference: install, then a wall of tables. It now
teaches the loop (quick start, what the hook does, asking questions,
reviewing a diff, keeping the index fresh, recovering raw output, tuning,
custom filters) before the reference, with troubleshooting at the end.
Every example output was captured from the real binary.

Per-agent guides in docs/agents/ document what install-hook writes, what is
intercepted, how to verify and how to remove, since each harness differs
(Copilot is the only one with PostToolUse, Codex's matcher excludes reads,
Antigravity answers with decision allow/deny, OpenCode is MCP-only).

Also corrects claims the code contradicts: Java/C#/Ruby/PHP/Kotlin/Swift are
not indexed by default, repo secret-rules can only add, `filter generate`
sends redacted recordings to an AI CLI, and Copilot installs globally unless
--local.


* test: rename a trust-gate variable CodeQL reads as sensitive

rust/cleartext-logging flagged `trusted` interpolated into an assert
message as sensitive data. It holds `echo keep-line` output; the rename
removes the false positive instead of dismissing it.


* chore(deps): bump lru 0.18.0 -> 0.18.4 for RUSTSEC-2026-0253

LruCache::pop() is not panic-safe and can use-after-free (pulled in by
ratatui-core). RustSec classifies it as informational `unsound`, so
cargo-deny does not fail on it, but OpenSSF Scorecard reports it as an
open vulnerability. Lockfile-only, semver-compatible.


* chore(deps): refresh the lockfile and raise MSRV to 1.90

tree-sitter 0.27 already needs 1.90, so the declared 1.88 floor was false and
only held because CI builds on stable. getrandom moves to 0.4, the version
tempfile already pulls in, so the daemon token adds no compilation unit.
Checked with cargo +1.90.0 check --locked --all-targets.


* fix(filters): keep the signal agents act on and run Bash commands under bash

A benchmark against rtk/squeez and a Haiku 4.5 A/B session found filters
hiding exactly what an agent needs to fix its code:

- cargo build/check/clippy/test never filtered stderr, where rustc writes
  every diagnostic (0% reduction on clippy). They now set filter_stderr and
  a new block_caps rule keeps each warning's message and location; errors
  are never capped.
- git-log replaced hashes with <commit_hash>, leaving nothing to git show.
  It keeps 12-char hashes, with extract_sections split_on_start/max_lines.
- git-diff dropped @@ hunk headers, the only line numbers an agent edits by.
- node-test dropped the tests/pass counts and, behind max_lines 60 and
  repeated MODULE_TYPELESS warnings, the expected/actual values. Replayed
  over 24 real outputs: 0 of 247 signal lines lost, 67% fewer tokens.
- Seven bundled regexes never compiled (lookarounds, bare braces) and were
  skipped silently. regex_issues lints every field; a test gates the
  bundled corpus and doctor reports user/local filters.

On Windows, tokenix run executed the rewritten Bash-tool command under
cmd /C, so heredocs, $(...), 2>/dev/null and echo broke and the agent
burned turns on it. It now reuses the Git Bash it was launched from and
falls back to cmd only outside one. Unit tests stop reading the
developer's ~/.tokenix, whose filter copies shadowed the bundled ones.


* ci: give verify.sh's platform case a default branch

SonarCloud flags a case without a `*)` arm (shelldre:S131). Non-Windows
hosts already fell through; the arm makes that explicit.


---------

+++ src/chunker.rs
@@ -420,7 +420,7 @@ fn find_first_identifier<'a>(node: tree_sitter::Node<'a>, source: &'a [u8]) -> O
             return Some(text.to_string());
         }
     }
-    for i in 0..node.child_count() as u32 {
+    for i in 0..node.child_count() {
         if let Some(child) = node.child(i) {
             if let Some(name) = find_first_identifier(child, source) {
                 return Some(name);
@@ -471,7 +471,7 @@ fn chunk_with_parser(
                 kind: kind.to_string(),
             });
         }
-        for i in 0..node.child_count() as u32 {
+        for i in 0..node.child_count() {
             if let Some(child) = node.child(i) {
                 traverse(child, source, is_symbol_node, symbols);
             }
